ARTEMIS-4302 Renaming JournalTransactionForget as JournalTransactionForgetTest
Renaming JournalTransactionForgetTest so it is picked by maven plugin test filter.

As I'm fixing JournalTransactionForget test I also found an older test in the same situation and I'm fixing that one as well

(cherry picked from commit 6e36be67ae07d9ca1bf64e6c2410a45434a5168c)

downstream: ENTMQBR-8489
